refactor: designsystem utils 폴더 정리
