+++ core/common/src/main/java/com/teampatch/core/common/Activity.kt
@@ -3,6 +3,10 @@ package com.teampatch.core.common
 import android.app.Activity
 import android.content.Context
 import android.content.ContextWrapper
+import android.content.Intent
+import android.os.Build
+import android.provider.Settings
+import androidx.annotation.RequiresApi
 
 fun Context.findActivity(): Activity? {
     var context = this
@@ -12,4 +16,10 @@ fun Context.findActivity(): Activity? {
     }
 
     return null
+}
+
+fun Context.startNotificationSettingsActivity(packageName: String) {
+    val intent = Intent(Settings.ACTION_APP_NOTIFICATION_SETTINGS)
+        .putExtra(Settings.EXTRA_APP_PACKAGE, packageName)
+    startActivity(intent)
 }
