+++ core/common/src/main/java/com/teampatch/core/common/Paging.kt
@@ -1,0 +1,20 @@
+package com.teampatch.core.common
+
+import androidx.paging.LoadState
+import androidx.paging.LoadStates
+import androidx.paging.PagingData
+import androidx.paging.compose.LazyPagingItems
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flowOf
+
+fun <T: Any> Throwable.toPagingData(): Flow<PagingData<T>> {
+    val errorLoadStates = LoadStates(
+        refresh = LoadState.Error(this),
+        prepend = LoadState.Error(this),
+        append = LoadState.Error(this)
+    )
+    return flowOf(PagingData.empty(errorLoadStates))
+}
+
+fun<T: Any> LazyPagingItems<T>.getOrNull(index: Int): T? =
+    kotlin.runCatching { get(index) }.getOrNull()
