+++ core/designsystem/src/main/java/com/teampatch/core/designsystem/component/AdditionMemoryCard.kt
@@ -39,6 +39,7 @@ import com.teampatch.core.designsystem.theme.G4
 import com.teampatch.core.designsystem.theme.HarmonyTheme
 import com.teampatch.core.designsystem.theme.MainGreen
 import com.teampatch.core.designsystem.theme.PretendardFontFamily
+import com.teampatch.core.designsystem.utils.noRippleClickable
 
 @Composable
 fun ColumnScope.AdditionMemoryCard(
