+++ core/designsystem/src/main/java/com/teampatch/core/designsystem/component/DefaultButton.kt
@@ -24,6 +24,7 @@ import com.teampatch.core.designsystem.theme.HarmonyTheme
 import com.teampatch.core.designsystem.theme.MainGreen
 import com.teampatch.core.designsystem.theme.PretendardFontFamily
 import com.teampatch.core.designsystem.theme.WH
+import com.teampatch.core.designsystem.utils.noRippleClickable
 
 @Composable
 fun DefaultButton(
