+++ core/designsystem/src/main/java/com/teampatch/core/designsystem/component/RoundButton.kt
@@ -19,6 +19,7 @@ import com.teampatch.core.designsystem.theme.HarmonyTheme
 import com.teampatch.core.designsystem.theme.MainGreen
 import com.teampatch.core.designsystem.theme.PretendardFontFamily
 import com.teampatch.core.designsystem.theme.WH
+import com.teampatch.core.designsystem.utils.noRippleClickable
 
 @Composable
 fun RoundButton(
