+++ feature/question-expand/build.gradle.kts
@@ -11,6 +11,7 @@ android {
 
 dependencies {
 
+    implementation(project(":core:common"))
     implementation(project(":core:domain"))
     implementation(project(":core:designsystem"))
