+++ feature/question-expand/src/main/java/com/teampatch/feature/question/expand/QuestionExpandScreen.kt
@@ -29,16 +29,16 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.paging.PagingData
 import androidx.paging.compose.collectAsLazyPagingItems
+import com.teampatch.core.common.getOrNull
 import com.teampatch.core.designsystem.R.drawable.ic_chevron_question
 import com.teampatch.core.designsystem.component.BackButtonAppBar
-import com.teampatch.core.designsystem.component.noRippleClickable
-import com.teampatch.core.designsystem.getOrNull
 import com.teampatch.core.designsystem.theme.BL
 import com.teampatch.core.designsystem.theme.G1
 import com.teampatch.core.designsystem.theme.G3
 import com.teampatch.core.designsystem.theme.G4
 import com.teampatch.core.designsystem.theme.HarmonyTheme
 import com.teampatch.core.designsystem.theme.PretendardFontFamily
+import com.teampatch.core.designsystem.utils.noRippleClickable
 import com.teampatch.core.domain.fake.FakeQuestions
 import com.teampatch.feature.question.expand.model.QuestionExpandSideEffect
 import com.teampatch.feature.question.expand.model.QuestionExpandUiState
